Relacion 1:1 y nueva tabla usuarios

+++ app/Http/Controllers/borrarController.php
@@ -9,7 +9,14 @@ class borrarController extends Controller
 {
     public function borrar($id){
 
-        $post = Post::findorfail($id);
+        $post = Post::find($id);
+
+        if(!$post)
+        {
+            abort(404);
+        }
+        $usuario = $post->user;
+        $usuario->delete();
         $post->delete();
         return redirect()->route('inicio');
     }
